Crate sda_client

This crate provides the basic functionality and abstract logic typically needed by participants, clerks, and recipients for interacting with an SDA service.

Structs

EncryptionKeypair
Description of an encryption keypair.

ParticipantInput
Typed input to be turned into a participation.

SdaClient
Primary object for interacting with the SDA service.

SdaClientError
The Error type.

SignatureKeypair
Description of a signature keypair.

Traits

Clerking
Basic tasks typically performed by a clerk.

KeyStorage
Trait for accessing keys stored in keystore.

Keystore
Requirements for any keystore used by the client.

Maintenance
Basic tasks typically performed with respect to agent associated resources.

Participating
Basic tasks typically performed by a participant.

Receiving
Basic tasks typically performed by a recipient.

Type Definitions

SdaClientResult
Convenient wrapper around std::Result.