Struct sda_client::SdaClient

pub struct SdaClient {
    pub agent: Agent,
    // some fields omitted
}

Primary object for interacting with the SDA service.

For instance used by participants to provide input to aggregations and by clerks to perform their clerking tasks.

Fields
agent: Agent

Agent to be used when e.g. identifying with the service.

Methods

impl SdaClient

fn new_agent(keystore: Arc<Keystore>) -> SdaClientResult<Agent>

Create a new agent, including mandatory signature keypair.

impl SdaClient

fn new(agent: Agent,
keystore: Arc<Keystore>,
service: Arc<SdaService>)
-> SdaClient

Create a new client.