Trait sda_client::Clerking

pub trait Clerking {
    fn clerk_once(&self) -> SdaClientResult<bool>;
    fn run_chores(&self, max_iterations: isize) -> SdaClientResult<()>;
}

Basic tasks typically performed by a clerk.

Required Methods

fn clerk_once(&self) -> SdaClientResult<bool>

Execute clerking process once: download, process, and upload the next job pending on the service, if any.

fn run_chores(&self, max_iterations: isize) -> SdaClientResult<()>

This is the typical method clerks should use. Execute routine chores, including registration and clerking.

Note that a negative max_iterations will continue the clerking process until the service reports that there are no more jobs.